+++ packages/xchain-bitcoin/src/utils.ts
@@ -154,6 +154,7 @@ export const scanUTXOs = async ({
   network,
   address,
   confirmedOnly = true, // default: scan only confirmed UTXOs
+  fetchTxHex,
 }: ScanUTXOParam): Promise<UTXO[]> => {
   switch (network) {
     case Network.Testnet: {
@@ -171,18 +172,27 @@ export const scanUTXOs = async ({
         utxos = await sochain.getUnspentTxs(addressParam)
       }
 
-      return utxos.map(
-        (utxo) =>
-          ({
-            hash: utxo.txid,
-            index: utxo.output_no,
+      const results: UTXO[] = []
+
+      for (const utxo of utxos) {
+        let txHex
+        if (fetchTxHex) {
+          txHex = (await sochain.getTx({ hash: utxo.txid, sochainUrl, network })).tx_hex
+        }
+
+        results.push({
+          hash: utxo.txid,
+          index: utxo.output_no,
+          value: assetToBase(assetAmount(utxo.value, BTC_DECIMAL)).amount().toNumber(),
+          witnessUtxo: {
             value: assetToBase(assetAmount(utxo.value, BTC_DECIMAL)).amount().toNumber(),
-            witnessUtxo: {
-              value: assetToBase(assetAmount(utxo.value, BTC_DECIMAL)).amount().toNumber(),
-              script: Buffer.from(utxo.script_hex, 'hex'),
-            },
-          } as UTXO),
-      )
+            script: Buffer.from(utxo.script_hex, 'hex'),
+          },
+          txHex,
+        })
+      }
+
+      return results
     }
     case Network.Mainnet:
     case Network.Stagenet: {
@@ -225,6 +235,7 @@ export const buildTx = async ({
   sochainUrl,
   haskoinUrl,
   spendPendingUTXO = false, // default: prevent spending uncomfirmed UTXOs
+  fetchTxHex = false,
 }: TxParams & {
   feeRate: FeeRate
   sender: Address
@@ -232,10 +243,11 @@ export const buildTx = async ({
   sochainUrl: string
   haskoinUrl: string
   spendPendingUTXO?: boolean
-}): Promise<{ psbt: Bitcoin.Psbt; utxos: UTXO[] }> => {
+  fetchTxHex?: boolean
+}): Promise<{ psbt: Bitcoin.Psbt; utxos: UTXO[]; inputs: UTXO[] }> => {
   // search only confirmed UTXOs if pending UTXO is not allowed
   const confirmedOnly = !spendPendingUTXO
-  const utxos = await scanUTXOs({ sochainUrl, haskoinUrl, network, address: sender, confirmedOnly })
+  const utxos = await scanUTXOs({ sochainUrl, haskoinUrl, network, address: sender, confirmedOnly, fetchTxHex })
 
   if (utxos.length === 0) throw new Error('No utxos to send')
   if (!validateAddress(recipient, network)) throw new Error('Invalid address')
@@ -287,7 +299,7 @@ export const buildTx = async ({
     }
   })
 
-  return { psbt, utxos }
+  return { psbt, utxos, inputs }
 }
 
 /**